Implement control logic for task 1 (#86)
* Small syntax/logic fixes

* Add searching parameters to config.yaml

* Generate searching path for hotspots

* Parse config values properly

* Fix typo

* Fix drone flying in a circle

In `generate_circular_path`, the first point is appended to the returned
list of waypoints, so that the drone actually completes a full circle.

* Update tests for plot_circular_path.py

* Remove positive altitude constraint

Probably due to changed library, it is no longer necessary to check that the relative altitude is above zero.
Evan Zhao wrote this code a while ago, I can't really remember why I checked for positive relative altitude XD

* Address recommended changes

* Change config parameters

Changed
* Drone takeoff altitude (40 -> 10 meters)
* Drone timeout, because uploading commands can take a while (30 -> 120
  seconds)